+++ packages/blocks/blocksBasic/webpack.demo.js
@@ -1,12 +1,15 @@
 const CopyPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
+const type = require('@lowdefy/helpers').type;
 const path = require('path');
 const { merge } = require('webpack-merge');
 
 const common = require('./webpack.common.js');
 const packageJson = require('./package.json');
 
-const port = process.argv[process.argv.findIndex((val) => val === '--port') + 1] || 3002;
+const port = type.isNumber(process.argv[process.argv.findIndex((val) => val === '--port') + 1])
+  ? process.argv[process.argv.findIndex((val) => val === '--port') + 1]
+  : 3002;
 
 const sanitizeName = (name) => {
   return name