+++ packages/nunjucks/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/lowdefy/lowdefy/compare/@lowdefy/nunjucks@1.1.0...@lowdefy/nunjucks@1.1.2) (2020-12-15)
+
+**Note:** Version bump only for package @lowdefy/nunjucks
+
+
+
+
+
+## [1.1.1](https://github.com/lowdefy/lowdefy/compare/@lowdefy/nunjucks@1.1.0...@lowdefy/nunjucks@1.1.1) (2020-12-15)
+
+**Note:** Version bump only for package @lowdefy/nunjucks
+
+
+
+
+
 # 1.1.0 (2020-12-10)