+++ packages/servers/serverDocker/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1](https://github.com/lowdefy/lowdefy/compare/@lowdefy/server-docker@0.0.0-alpha.7...@lowdefy/server-docker@0.0.1) (2020-12-15)
+
+**Note:** Version bump only for package @lowdefy/server-docker
+
+
+
+
+
+# [0.0.0](https://github.com/lowdefy/lowdefy/compare/@lowdefy/server-docker@0.0.0-alpha.7...@lowdefy/server-docker@0.0.0) (2020-12-15)
+
+**Note:** Version bump only for package @lowdefy/server-docker
+
+
+
+
+
 # 0.0.0-alpha.7 (2020-12-10)